PieChart: follow-up bugfix
Making sorting of slices by value configurable introduced an issue
with small values. The original raphael pie implementation assumed
the values were always sorted.

Now, if clientSorting is disabled, small slices are not accumulated and
merged into a "Others" category. This should probably be implemented
properly in the future for consistent behaviour.